+++ src/renderer/preload.d.ts
@@ -90,6 +90,7 @@ declare global {
         ): void;
 
         clipVideo(video: RendererVideo, offset: number, duration: number): void;
+        getHardwareAcceleration(): boolean;
       };
     };
   }